% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.cirwin.R
\name{area.cirwin}
\alias{area.cirwin}
\title{Area of a Window}
\usage{
\method{area}{cirwin}(W)
}
\arguments{
\item{W}{A window, whose area will be computed.
This should be an object of class \code{\link{cirwin}}.}
}
\value{
A numerical value giving the area of the window
}
\description{
Computes the area of a window
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
