% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirwin.R
\name{cirwin}
\alias{cirwin}
\title{Create a window on the circle}
\usage{
cirwin(type = "circle", intervals)
}
\arguments{
\item{type}{The type of window to be created. 
Can be either the full circle or a subset and should be "\code{circle}" respectively "\code{arcs}".
If type is "\code{circle}" the second argument \code{intervals} will be ignored}

\item{intervals}{Only required for \code{type="arcs"}. 
A vector or matrix like object with length 2 respectively number of columns equal to \code{2*n}, where \code{n} is some positive integer
If specified as a vector (or matrix) the entries (columns) should be alternating between lower and upper bound, starting with the lower bound.}
}
\value{
An object of class \code{cirwin}, which contains information regarding the window specified
}
\description{
Creates an object of class "\code{cirwin}" representing an observation window on the circle
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
