% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_uni_vM_mix.R
\name{d_uni_vM_mix}
\alias{d_uni_vM_mix}
\title{Mixture Density Function of Uniform and von Mises}
\usage{
d_uni_vM_mix(x, mu, kappa, p_unif)
}
\arguments{
\item{x}{An object of class \code{\link{ppc}}.}

\item{mu}{Mean direction of the distribution. Is coerced to class \code{\link{circular}}.}

\item{kappa}{Non-negative numeric value for the concentration parameter of the distribution.}

\item{p_unif}{A numeric between 0 and 1. The probability of observing a point from the uniform distribution.}
}
\value{
The density at the locations given by \code{x}.
}
\description{
Density function for the mixture of the uniform and von Mises distribution.
}
