% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.cirwin.R
\name{inside.cirwin}
\alias{inside.cirwin}
\title{Test whether points are inside a window on a circle}
\usage{
inside.cirwin(points, W)
}
\arguments{
\item{points}{Vector of angles to be tested.}

\item{W}{Awindow. This should be an object of class \code{\link{cirwin}}.}
}
\value{
A logical vector whose \code{i}th entry is \code{TRUE} if the corresponding point \code{ang[i]} is inside \code{W}.
}
\description{
Test whether points lie inside or outside a given window of class \code{cirwin}
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
