% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcir.R
\name{Kcir}
\alias{Kcir}
\title{K-function of a three-dimensional point pattern}
\usage{
Kcir(X, ..., r = NULL, rmax = NULL, nrval = 128)
}
\arguments{
\item{X}{Circular point pattern (object of class \code{\link{ppc}}).}

\item{...}{Ignored.}

\item{r}{Optional.
Vector of values for the argument r at which \code{K(r)} should be evaluated.
Users are advised not to specify this argument; there is a sensible default.
If necessary, specify rmax.}

\item{rmax}{Optional. Maximum desired value of the argument r.}

\item{nrval}{Optional.
Number of values of r for which \code{K3(r)} will be estimated.
A large value of nrval is required to avoid discretisation effects.}
}
\value{
A function value table (object of class \code{\link{fv}}) that can be plotted, printed or coerced to a data frame containing the function values.
}
\description{
Estimates the K-function from a point pattern on the circle
}
\details{
No edge correction have been implemented.
And it is assumede that the data are observed on the entire circle, i.e. the window is the entire circle.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
