% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npoints.ppc.R
\name{npoints.ppc}
\alias{npoints.ppc}
\title{Number of points in a point pattern}
\usage{
\method{npoints}{ppc}(x)
}
\arguments{
\item{x}{A point pattern (object of class \code{\link{ppc}})}
}
\value{
Integer
}
\description{
Get the number of points in a point pattern of type \code{\link{ppc}}
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
