% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ppc.R
\name{plot.ppc}
\alias{plot.ppc}
\title{Plot \code{ppc} object}
\usage{
\method{plot}{ppc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ppc}}. Must be observed on intire circle.}

\item{...}{Additional arguments passed to \code{\link{plot.circular}}.}
}
\description{
Function for plotting an object of class ppc (Only implemented for data observed on the full circle)
}
