% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2car.R
\name{pol2car}
\alias{pol2car}
\title{Polar to cartesian coordinates}
\usage{
pol2car(X, deg = FALSE, rep = "std")
}
\arguments{
\item{X}{A numerical vector or matrix whos entries constitute polar coordinates.
If \code{X} is a matrix it should have two columns, consituting the angle and radial distance.
If \code{X} is a vector it is assumed that all entries are to be considered as angles.}

\item{deg}{Logical value indicating whether the input is in randians or degrees. Radian is default.}

\item{rep}{Determine the representation of the polar coordinates.
If \code{"std"} the angle will range from \code{0} to \code{2 * pi}.
The classical representation of polar coordinates in radians.
If \code{"neg"} the angle will range from \code{-pi} to \code{pi}.
Deafult is "std".}
}
\value{
A \code{data.frame} with 2 columns representing the cartesian coordinate set.
}
\description{
Function for transforming polar coordinates to cartesian coordinates
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
