% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc}
\alias{ppc}
\title{Create point pattern}
\usage{
ppc(x, y, window = cirwin(), check = TRUE)
}
\arguments{
\item{x}{Vector of polar angles or cartesian x coordinates of the data points.}

\item{y}{Vector of cartesian y coordinates. Should be specified only if \code{x} represent cartesian coordiante}

\item{window}{Window of observation, an object of class \code{\link{cirwin}}.}

\item{check}{Logical value indicating wheter to check that al points lie inside the specified window.}
}
\value{
An object of class \code{ppc} describing a point pattern on the unit circle.
}
\description{
Creates an object of class \code{ppc} representing a point pattern dataset on the unit circle.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
