% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cirwin.R
\name{print.cirwin}
\alias{print.cirwin}
\title{Print Brief Details of a Spatial Window}
\usage{
\method{print}{cirwin}(x, ...)
}
\arguments{
\item{x}{Window (object of class "\code{cirwin}")}

\item{...}{Ignored}
}
\value{
A very brief description of the window \code{x} is printed.
}
\description{
Prints a very brief description of a window object of class \code{\link{cirwin}}.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
