% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppc.R
\name{print.ppc}
\alias{print.ppc}
\title{Print Brief Details of a Point Pattern Dataset}
\usage{
\method{print}{ppc}(x, ...)
}
\arguments{
\item{x}{Point pattern on the unit circle (object of class \code{\link{ppc}})}

\item{...}{Ignored}
}
\value{
A very brief description of the point pattern \code{x} is printed.
}
\description{
Prints a very brief description of a point pattern dataset of class \code{\link{ppc}}.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
