% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cirwin.R
\name{summary.cirwin}
\alias{summary.cirwin}
\title{Summary of a Circular Window}
\usage{
\method{summary}{cirwin}(object, ...)
}
\arguments{
\item{object}{Window (object of class "cirwin").}

\item{...}{Ignored.}
}
\description{
Prints a useful description of a window object of class \code{\link{cirwin}}.
}
\details{
A useful description of the window object is printed.
This is a method for the generic function \code{\link{summary}}.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
